Move fflog publicly
